refactored validation
* introduced BaseValidator
* added ConnnectionValidator
* refactored directory structure of sparql directories
* added sparql files with checks of the invariants for message references
