+++ webofneeds/won-core/src/main/java/won/protocol/validation/WonSparqlValidator.java
@@ -1,4 +1,20 @@
-package won.protocol.message;
+/*
+ * Copyright 2012  Research Studios Austria Forschungsges.m.b.H.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package won.protocol.validation;
 
 import org.apache.jena.graph.Node;
 import org.apache.jena.query.*;
@@ -22,14 +38,14 @@ public class WonSparqlValidator {
 
     public WonSparqlValidator(Query constraint) {
         if (!constraint.isAskType() && !constraint.isSelectType()) {
-            throw new IllegalArgumentException("Wrong constriant type!");
+            throw new IllegalArgumentException("Wrong constraint type!");
         }
         this.constraint = constraint;
     }
 
     public WonSparqlValidator(Query constraint, String name) {
         if (!constraint.isAskType() && !constraint.isSelectType()) {
-            throw new IllegalArgumentException("Wrong constriant type!");
+            throw new IllegalArgumentException("Wrong constraint type!");
         }
         this.constraint = constraint;
         this.name = name;
@@ -47,13 +63,23 @@ public class WonSparqlValidator {
         return new ValidationResult(false, "Invalid constraint: " + constraint.toString());
     }
 
+
     private ValidationResult validateSelect(final Dataset input) {
         try (QueryExecution qe = QueryExecutionFactory.create(constraint, input)) {
+
             ResultSet result = qe.execSelect();
+            if (!result.hasNext()) {
+                //this is a valid result if the projection vars don't contain 'check' (in which case we want exactly one result, see below)
+                if (constraint.getProjectVars().stream().noneMatch(var -> "check".equals(var.getVarName()))) {
+                    //no 'check' variable: we have a valid result.
+                    return new ValidationResult();
+                }
+            }
             while (result.hasNext()) {
                 Binding binding = result.nextBinding();
                 Node node = binding.get(SELECT_VALIDATION_VARIABLE);
                 if (node != null) {
+                    //there is a binding for a variable with name 'check': check its value:
                     if (node.isLiteral()) {
                         String resultString = node.getLiteralValue().toString();
                         if (SELECT_VALIDATION_PASSED_VALUE.equals(resultString)) {
@@ -63,9 +89,21 @@ public class WonSparqlValidator {
                                     .toString());
                         }
                     }
+                } else {
+                    // there is no binding for a variable with name 'check':
+                    // in this case, we do it similar to checking with ASK: if there are solutions,
+                    // they reveal violations of the validity checks
+                    // ...
+                    // in order to keep results small, we only report the first binding in the ValidationResult
+                    String errorMessage = "SPARQL query produced this solution, which indicates a problem: " + binding.toString() + ", query: " + constraint.toString(Syntax.syntaxSPARQL_11);
+                    if (result.hasNext()) {
+                        //just inform that there are more results
+                        errorMessage += ". Note: this is only the first solution. There are more problems.";
+                    }
+                    return new ValidationResult(false, errorMessage);
                 }
             }
-            return new ValidationResult(false, "SELECT query did not produce a binding for required variable 'check'");
+            throw new IllegalStateException("We should have returned a result earlier. Constraint: " + constraint.toString(Syntax.syntaxSPARQL_11));
         }
     }
