+++ webofneeds/won-core/src/main/resources/validation/connection/connect_message_has_no_previous_messages.rq
@@ -1,0 +1,7 @@
+# The `Create` message of a need must not contain a reference
+prefix msg: <http://purl.org/webofneeds/message#>
+select ?msg where {
+  ?msg msg:hasMessageType msg:CreateMessage;
+	   msg:hasPreviousMessage ?msg.
+}
+
