+++ webofneeds/won-core/src/main/resources/validation/connection/path_to_create_message.rq
@@ -1,0 +1,11 @@
+# From any message, other than a `Create` message, there must be
+# at least one path to the need's `Create` message
+PREFIX msg: <http://purl.org/webofneeds/message#>
+SELECT ?msg WHERE {
+  ?msg msg:hasMessageType ?msgType.
+  FILTER NOT EXISTS {
+    ?msg msg:hasPreviousMessage* ?createMsg.
+    ?createMsg msg:hasMessageType msg:CreateMessage.
+  }
+  FILTER (?msgType != msg:NeedCreatedNotificationMessage)
+}
