+++ webofneeds/won-core/src/main/resources/validation/connection/reference_in_response.rq
@@ -1,0 +1,22 @@
+# A Response must always reference the message that it is a response to.
+PREFIX msg: <http://purl.org/webofneeds/message#>
+SELECT * WHERE {
+  {
+  	?resp a msg:FromSystem .
+	?resp msg:isResponseTo ?msg .
+  } UNION {
+	?resp a msg:FromOwner .
+	?resp msg:isResponseTo ?msg .
+  } UNION {
+	?resp a msg:FromExternal .
+	?resp msg:isRemoteResponseTo ?msg .
+  }
+  {
+  	?resp msg:hasMessageType msg:SuccessResponse.
+  } UNION {
+    ?resp msg:hasMessageType msg:FailureResponse.
+  }
+  FILTER NOT EXISTS {
+  	?resp msg:hasPreviousMessage ?msg.
+  }
+}
