+++ webofneeds/won-core/src/main/resources/validation/message/01_basic/invalid_default_graph.rq
