+++ webofneeds/won-core/src/main/resources/validation/message/02_prop/invalid_direction.rq
