+++ webofneeds/won-core/src/main/resources/validation/message/02_prop/missing_direction.rq
