+++ webofneeds/won-core/src/main/resources/validation/message/03_chain/invalid_content_chain.rq
