+++ webofneeds/won-core/src/main/resources/validation/message/04_uri/graph_uris.rq
