+++ webofneeds/won-core/src/main/resources/validation/message/05_sign/signature_properties.rq
