+++ webofneeds/won-core/src/main/resources/validation/message/05_sign/signed_or_signature.rq
