+++ webofneeds/won-core/src/main/resources/validation/message/06_types/invalid_create_msg_prop.rq
